+++ lib/util/util.h
@@ -98,7 +98,7 @@ int memcmp_const(const void *a, const void *b, size_t len);
 
 
 // size is the size of the input string, must be even
-void hex2bin(const char *in, char* out, size_t size);
+void hex2bin(const char *in, char* out, int size);
 
 void print_hex(const char *buf, int size);
